atr_bot/atr_bot/requirements.txt
ccxt==4.3.68
requests==2.32.3
PyYAML==6.0.2
pandas==2.2.3
python-dotenv==1.0.1
tenacity==8.3.0
